Pulling request wrapper into separate class
